Refactor the alpha_vantage menu to replace the market_view commands in fa
