+++ gamestonk_terminal/fundamental_analysis/av_view.py
